Separate domUtils into per category files
